% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classdistr.R
\name{classify_distribution}
\alias{classify_distribution}
\title{Classified observed data into a distribution class.}
\usage{
classify_distribution(data, n_bootstrap = 1000, pc_bootstrap = 0.9)
}
\arguments{
\item{n_bootstrap}{Number of bootstap iterations}

\item{x}{A numeric vector of observations}
}
\description{
Using bootstrapping to estimate likeliness, this function classifies
observed data into one of the following: Binary, Uniform, Skewed, Normal or Bimodal.
}
\examples{
set.seed(1234)

# Uniform distribution test
x <- runif(100)
classify_distribution(x)

# Normal distribution tests
x <- rnorm(100)
classify_distribution(x)

# Skewed data
x <- c(1, 2, 2, 2, 3, 4, 5, 5, 5, 6, 7, 8, 9, 10, 20, 30, 50, 100)
classify_distribution(x)

x <- c(1, 2, 2, 2, 3, 4, 5, 5, 5, 6, 7, 8, 9, 10, 20, 30, 50, 100, 1000000)
classify_distribution(x)

x <- rlnorm(100)
classify_distribution(x)

x <- exp(1:100)
classify_distribution(x)

x <- rpois(100, lambda = 0.5)
classify_distribution(x)

x <- rweibull(100, shape = 0.5)
classify_distribution(x)

### Bimodal data NOT RUN
# library(truncnorm)
# nn <- 1e4
# sims <- c(rtruncnorm(nn/2, a=1, b=5, mean=2, sd=.5),
#          rtruncnorm(nn/2, a=1, b=5, mean=4, sd=.5))
# classify_distribution(sims)
}
