% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classdistR6.R
\name{distrnorm}
\alias{distrnorm}
\title{Creates a recommended classInt based on the type of distribution.}
\description{
Creates a recommended classInt based on the type of distribution.

Creates a recommended classInt based on the type of distribution.
}
\details{
Creates a distrnorm R6 class for recommending a classInt based on the shape
of the distribution of the observed data
}
\examples{
set.seed(1234)

# Binary distribution test
x <- sample(c(0,1), 100, replace = TRUE)
distrnorm$new(x)
# Uniform distribution test
x <- runif(100)
distrnorm$new(x)

# Normal distribution tests
x <- rnorm(100)
distrnorm$new(x)

# Skewed data
x <- c(1, 2, 2, 2, 3, 4, 5, 5, 5, 6, 7, 8, 9, 10, 20, 30, 50, 100)
distrnorm$new(x)

x <- c(1, 2, 2, 2, 3, 4, 5, 5, 5, 6, 7, 8, 9, 10, 20, 30, 50, 100, 1000000)
distrnorm$new(x)

x <- rlnorm(100)
distrnorm$new(x)

x <- exp(1:100)
distrnorm$new(x)

x <- rpois(100, lambda = 0.5)
distrnorm$new(x)

x <- rweibull(100, shape = 0.5)
distrnorm$new(x)

### Bimodal data NOT RUN
# library(truncnorm)
# nn <- 1e4
# sims <- c(rtruncnorm(nn/2, a=1, b=5, mean=2, sd=.5),
#          rtruncnorm(nn/2, a=1, b=5, mean=4, sd=.5))
# distrnorm$new(sims)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{(\code{numeric()})\cr
Original observations}

\item{\code{outliers}}{(\code{logical()})\cr
Logical vector indicating is observations are
outliers}

\item{\code{fitted_distribution}}{(\code{character()})\cr
Suggested distribution}

\item{\code{normalisation}}{(\code{character()})\cr
Recommended class interval style based on
distribution}

\item{\code{breaks}}{(\code{numeric()})\cr
Recommended breaks for classes}

\item{\code{number_of_classes}}{(\code{numeric()})\cr
Number of classes identified}

\item{\code{normalised_data}}{(\code{numeric()})\cr
Normalised values based on recommendations}

\item{\code{polarity}}{(\code{numeric(1)})\cr
Which direction should the normalisation occur}

\item{\code{percentiles}}{(\code{numeric()})\cr
Observation percentiles}

\item{\code{fittedmodel}}{(\code{character()})\cr
Fitted univariate model}

\item{\code{model}}{(\code{univariateML()})\cr
Fitted univariate model parameters}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-distrnorm-new}{\code{distrnorm$new()}}
\item \href{#method-distrnorm-print}{\code{distrnorm$print()}}
\item \href{#method-distrnorm-plot}{\code{distrnorm$plot()}}
\item \href{#method-distrnorm-hist}{\code{distrnorm$hist()}}
\item \href{#method-distrnorm-setManualBreaks}{\code{distrnorm$setManualBreaks()}}
\item \href{#method-distrnorm-applyto}{\code{distrnorm$applyto()}}
\item \href{#method-distrnorm-to_data_frame}{\code{distrnorm$to_data_frame()}}
\item \href{#method-distrnorm-clone}{\code{distrnorm$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-distrnorm-new"></a>}}
\if{latex}{\out{\hypertarget{method-distrnorm-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6}
class.


Create a new distrnorm object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{distrnorm$new(
  x,
  polarity = 1,
  classint_preference = "jenks",
  num_classes_pref = NULL,
  potential_distrs = c("unif", "power", "norm", "lnorm", "weibull", "pareto", "exp")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A numeric vector of observations}

\item{\code{polarity}}{Which direction should the normalisation occur, defaults to
1 but can either be:
\itemize{
\item \strong{1:}: Lowest value is normalised to 0, highest value is
normalised to 1
\item \strong{-1:} Highest value is normalised to 0, lowest value is
normalised to 1
}}

\item{\code{classint_preference}}{Prefernce for classInt breaks}

\item{\code{num_classes_pref}}{Preference for number of classes for
classInt intervals}

\item{\code{potential_distrs}}{The types of distributions to fit}

\item{\code{n_bootstrap}}{Number of bootstrap iterations}

\item{\code{pc_bootstrap}}{Sampling proportion for bootstrapping}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{distrnorm} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-distrnorm-print"></a>}}
\if{latex}{\out{\hypertarget{method-distrnorm-print}{}}}
\subsection{Method \code{print()}}{
Prints the distrnorm
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{distrnorm$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-distrnorm-plot"></a>}}
\if{latex}{\out{\hypertarget{method-distrnorm-plot}{}}}
\subsection{Method \code{plot()}}{
Plots the normalised values against the original
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{distrnorm$plot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-distrnorm-hist"></a>}}
\if{latex}{\out{\hypertarget{method-distrnorm-hist}{}}}
\subsection{Method \code{hist()}}{
Histogram of normalised values against the
original
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{distrnorm$hist()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-distrnorm-setManualBreaks"></a>}}
\if{latex}{\out{\hypertarget{method-distrnorm-setManualBreaks}{}}}
\subsection{Method \code{setManualBreaks()}}{
Allows user to set manual breaks
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{distrnorm$setManualBreaks(brks)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{brks}}{User Defined Breaks}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-distrnorm-applyto"></a>}}
\if{latex}{\out{\hypertarget{method-distrnorm-applyto}{}}}
\subsection{Method \code{applyto()}}{
Applies the normalisation model to new data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{distrnorm$applyto(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A numeric vector of observations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-distrnorm-to_data_frame"></a>}}
\if{latex}{\out{\hypertarget{method-distrnorm-to_data_frame}{}}}
\subsection{Method \code{to_data_frame()}}{
Returns a data frame of the normalisation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{distrnorm$to_data_frame()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-distrnorm-clone"></a>}}
\if{latex}{\out{\hypertarget{method-distrnorm-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{distrnorm$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
