% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classdistr.R
\name{.check_for_outliers}
\alias{.check_for_outliers}
\title{Helper function to check for outliers}
\usage{
.check_for_outliers(data)
}
\arguments{
\item{data}{Observed data}
}
\description{
Helper function to check for outliers
}
\keyword{internal}
