% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classdistR6.R
\name{classdistr}
\alias{classdistr}
\title{Creates a recommended classInt based on the type of distribution.}
\description{
Creates a recommended classInt based on the type of distribution.

Creates a recommended classInt based on the type of distribution.
}
\details{
Creates a classdistr R6 class for recommending a classInt based on the shape of the distribution of the observed data
}
\examples{
set.seed(1234)

# Binary distribution test
x <- sample(c(0,1), 100, replace = TRUE)
classdistr$new(x)
# Uniform distribution test
x <- runif(100)
classdistr$new(x)

# Normal distribution tests
x <- rnorm(100)
classdistr$new(x)

# Skewed data
x <- c(1, 2, 2, 2, 3, 4, 5, 5, 5, 6, 7, 8, 9, 10, 20, 30, 50, 100)
classdistr$new(x)

x <- c(1, 2, 2, 2, 3, 4, 5, 5, 5, 6, 7, 8, 9, 10, 20, 30, 50, 100, 1000000)
classdistr$new(x)

x <- rlnorm(100)
classdistr$new(x)

x <- exp(1:100)
classdistr$new(x)

x <- rpois(100, lambda = 0.5)
classdistr$new(x)

x <- rweibull(100, shape = 0.5)
classdistr$new(x)

### Bimodal data NOT RUN
# library(truncnorm)
# nn <- 1e4
# sims <- c(rtruncnorm(nn/2, a=1, b=5, mean=2, sd=.5),
#          rtruncnorm(nn/2, a=1, b=5, mean=4, sd=.5))
# classdistr$new(sims)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{(\code{numeric()})\cr
Original observations}

\item{\code{outliers}}{(\code{logical()})\cr
Logical vector indicating is observations are outliers}

\item{\code{likely_distribution}}{(\code{character()})\cr
Likely distribution type including outliers}

\item{\code{recommended_normalisation}}{(\code{character()})\cr
Recommended class intervals based on distribution}

\item{\code{recommended_breaks}}{(\code{numeric()})\cr
Recommended breaks for classes}

\item{\code{number_of_classes}}{(\code{numeric()})\cr
Number of classes identified}

\item{\code{normalised_values}}{(\code{numeric()})\cr
Normalised values based on recommendations}

\item{\code{polarity}}{(\code{numeric(1)})\cr
Which direction should the normalisation occur}

\item{\code{percentiles}}{(\code{numeric()})\cr
Observation percentiles}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-classdistr-new}{\code{classdistr$new()}}
\item \href{#method-classdistr-print}{\code{classdistr$print()}}
\item \href{#method-classdistr-plot}{\code{classdistr$plot()}}
\item \href{#method-classdistr-clone}{\code{classdistr$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-classdistr-new"></a>}}
\if{latex}{\out{\hypertarget{method-classdistr-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.


Create a new classdistr object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{classdistr$new(
  x,
  polarity = 1,
  n_bootstrap = 20,
  pc_bootstrap = 0.7,
  classInt_preference = "jenks",
  num_classes_pref = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A numeric vector of observations}

\item{\code{polarity}}{Which direction should the normalisation occur, defaults to
1 but can either be:
\itemize{
\item \strong{1:}: Lowest value is normalised to 0, highest value is
normalised to 1
\item \strong{-1:} Highest value is normalised to 0, lowest value is
normalised to 1
}}

\item{\code{n_bootstrap}}{Number of bootstrap iterations}

\item{\code{pc_bootstrap}}{Sampling proportion for bootstrapping}

\item{\code{classInt_preference}}{Prefernce for classInt breaks}

\item{\code{num_classes_pref}}{Preference for number of classes for classInt intervals}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{classdistr} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-classdistr-print"></a>}}
\if{latex}{\out{\hypertarget{method-classdistr-print}{}}}
\subsection{Method \code{print()}}{
Prints the classdistr
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{classdistr$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-classdistr-plot"></a>}}
\if{latex}{\out{\hypertarget{method-classdistr-plot}{}}}
\subsection{Method \code{plot()}}{
Plots the normalised values against the original
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{classdistr$plot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-classdistr-clone"></a>}}
\if{latex}{\out{\hypertarget{method-classdistr-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{classdistr$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
