% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classdistr.R
\name{.classify_distribution}
\alias{.classify_distribution}
\title{Classified observed data into a distribution class.}
\usage{
.classify_distribution(data, n_bootstrap = 20, pc_bootstrap = 0.7)
}
\arguments{
\item{n_bootstrap}{Number of bootstrap iterations}

\item{pc_bootstrap}{Sampling proportion for bootstrapping}

\item{x}{A numeric vector of observations}
}
\description{
Based on a series of statistical tests, uses bootstrapping to classify
observed data into one of the following distributions: Binary, Uniform,
Skewed, Normal or Bimodal.
}
\keyword{internal}
