% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classdistr.R
\name{.classify_sample}
\alias{.classify_sample}
\title{Helper function to classify a single sample}
\usage{
.classify_sample(sample)
}
\arguments{
\item{sample}{Bootstrap sample}
}
\description{
Helper function to classify a single sample
}
\keyword{internal}
